% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcodeUnifiedPH.R
\name{UnifiedPH.EM}
\alias{UnifiedPH.EM}
\title{EM algorithm for arbitrarily censored data subject to left-truncation under the proportional hazards model}
\usage{
UnifiedPH.EM(
  d1,
  d2,
  d3,
  Li,
  Ri,
  Ei,
  Xp,
  n.int,
  order,
  g0,
  b0,
  tol,
  t.seq,
  equal = FALSE
)
}
\arguments{
\item{d1}{vector indicating whether an observation is exactly observed (1) or not (0).}

\item{d2}{vector indicating whether an observation is interval-censored (1) or not (0).}

\item{d3}{vector indicating whether an observation is right-censored (1) or not (0).}

\item{Li}{the left endpoint of the observed interval, if an observation is left-censored its corresponding entry should be 0.}

\item{Ri}{the right endpoint of the observed interval, if an observation is right-censored its corresponding entry should be Inf.}

\item{Ei}{the vector specifying the enrollment times, if no enrollment criteria is used then its corresponding entry should be 0.}

\item{Xp}{design matrix of predictor variables (in columns), should be specified without an intercept term.}

\item{n.int}{the number of interior knots to be used.}

\item{order}{the order of the basis functions.}

\item{g0}{initial estimate of the spline coefficients; should be of length n.int+order.}

\item{b0}{initial estimate of regression coefficients; should be of length dim(Xp)\link{2}.}

\item{tol}{the convergence criterion of the EM algorithm, see details for further description.}

\item{t.seq}{an increasing sequence of points at which the cumulative baseline hazard function is evaluated.}

\item{equal}{logical, if TRUE knots are spaced evenly across the range of the endpoints of the observed intervals and if FALSE knots are placed at quantiles. Defaults to FALSE.}
}
\value{
b: estimates of the regression coefficients.

g: estimates of the spline coefficients.

ll: the value of the maximized log-likelihood.

AIC: the Akaike information criterion.

BIC: the Bayesian information/Schwarz criterion.

bRi: I-spline basis matrix of dimension c(n.int+order, length(Ri)).

bLi: I-spline basis matrix of dimension c(n.int+order, length(Li)).

bt: I-spline basis matrix evaluated at the points t.seq.

mRi: M-spline basis matrix of dimension c(n.int+order, length(Ri)).

OPG: the variance covariance matrix of b and g
}
\description{
Fits a parametric proportional hazards model (PH), proposed in Gamage et al. (2022+), to arbitrarily censored data subject to left-truncation via an EM algorithm.
}
\details{
The above function fits a parametric proportional hazards model (PH), proposed in Gamage et al. (2022+), to arbitrarily censored data subject to left-truncation via an EM algorithm. For a discussion of starting values, number of interior knots, order, and further details please see Gamage et al. (2022+). The EM algorithm converges when the maximum absolute difference between consecutive parameter updates was less than the specified tol.
}
\references{
{
Withana Gamage, P., McMahan, C., and Wang, L. (2022+).
\emph{A Flexible Parametric Model for Fitting the Proportional Hazards Model}.
Submitted.
}
}
